I wanted to say thanks for creating this tool!
I wanted to tell you how much I appreciate that you took the time to create this tool. You are correct in saying that there are close to zero tools out there that provide this sort of conversion for Digital Forensics. I am taking a class and I needed something that would convert 128-bit SYSTEMTIME to human readable.
So again, thank you for putting all the time into this it looks great and works perfectly (for what I needed.)
-Mr. Hobbits